moves landing page content to the quizzes page and adds a base layout for the new landing page
