+++ index.html
@@ -16,6 +16,5 @@
     </header>
     <div class="app"></div>
     <script type="module" src="/main.js"></script>
-    <script type="module" src="/router.js"></script>
   </body>
 </html>
